small modify from exp6

+++ src/run_bert.py
@@ -44,7 +44,7 @@ def main(config, dirpath):
     best_model_path_list = []
     for i, (train_loader, valid_loader, weight) in enumerate(zip(train_loader_list, valid_loader_list, weight_list)):
         total_steps = epoch * len(train_loader)
-        warmup_steps = warmup_rate * total_steps
+        warmup_steps = int(warmup_rate * total_steps)
         model = LitBertForSequenceClassification(**config["network"], dirpath=dirpath, fold_id=i, weight=weight, num_warmup_steps=warmup_steps, num_training_steps=total_steps)
         checkpoint = pl.callbacks.ModelCheckpoint(monitor=f'valid_loss{i}', mode='min', save_top_k=1, save_weights_only=True, dirpath=dirpath, filename=f"fold{i}" + "{epoch}-{step}.ckpt")
         trainer = pl.Trainer(accelerator="gpu", devices=[gpu], max_epochs=epoch, gradient_clip_val=gradient_clip_val, callbacks=[checkpoint], logger=comet_logger)
